@@ -5,7 +5,6 @@ import { MainViewComponent } from './main-view.component';
 import { AuthGuardService } from '../Service/auth-guard.service';
 
 const routes: Routes = [
-  
    { path: '', component: MainViewComponent,
    children: [
      // tslint:disable-next-line: max-line-length
@@ -18,8 +17,9 @@ const routes: Routes = [
      { path: 'animalOverview', loadChildren: () => import('../animal-overview/animal-overview.module').then(m => m.AnimalOverviewModule), canActivate: [AuthGuardService] },
      // tslint:disable-next-line: max-line-length
      { path: 'animalProfile', loadChildren: () => import('../animal-profile/animal-profile.module').then(m => m.AnimalProfileModule), canActivate: [AuthGuardService] },
+     // tslint:disable-next-line: max-line-length
+     { path: 'animalMarket', loadChildren: () => import('../animal-market/animal-market.module').then(m => m.AnimalMarketModule), canActivate: [AuthGuardService] }
    ] }
-
 ];
 
 @NgModule({
